Rename old pyquaticus bridge
